optimized api structure
